@@ -1,30 +1,49 @@
 module vincit.fi/image-sorter
 
-go 1.14
+go 1.18
 
 require (
-	github.com/AllenDang/giu v0.5.6
-	github.com/AllenDang/imgui-go v1.12.1-0.20210801050644-e183d1c04c13 // indirect
+	github.com/AllenDang/giu v0.6.2
 	github.com/AndreasAbdi/gochromecast v0.0.0-20181018034447-700dddc0dea7
 	github.com/OpenDiablo2/dialog v0.0.0-20201230220514-26162241209f
 	github.com/disintegration/imaging v1.6.2
-	github.com/fatih/structs v1.1.0 // indirect
-	github.com/go-gl/glfw/v3.3/glfw v0.0.0-20210410170116-ea3d685f79fb
-	github.com/gogo/protobuf v1.3.1 // indirect
-	github.com/google/go-cmp v0.5.6 // indirect
+	github.com/go-gl/glfw/v3.3/glfw v0.0.0-20220320163800-277f93cfa958
 	github.com/google/uuid v1.1.1
 	github.com/hashicorp/mdns v1.0.3
-	github.com/imdario/mergo v0.3.9 // indirect
-	github.com/imroc/req v0.3.0 // indirect
 	github.com/nfnt/resize v0.0.0-20180221191011-83c6a9932646
 	github.com/pixiv/go-libjpeg v0.0.0-20190822045933-3da21a74767d
 	// Using customized version. Leaving this here to make sure the
 	// credit goes to correct person
 	//github.com/rivo/duplo v0.0.0-20180323201418-c4ec823d58cd
 	github.com/rwcarlsen/goexif v0.0.0-20190401172101-9e8deecbddbd
-	github.com/stretchr/testify v1.7.0
+	github.com/stretchr/testify v1.7.1
 	github.com/upper/db/v4 v4.0.1
 	github.com/vardius/message-bus v1.1.4
-	golang.org/x/image v0.0.0-20200618115811-c13761719519 // indirect
+)
+
+require (
+	github.com/AllenDang/go-findfont v0.0.0-20200702051237-9f180485aeb8 // indirect
+	github.com/AllenDang/imgui-go v1.12.1-0.20220322114136-499bbf6a42ad // indirect
+	github.com/TheTitanrain/w32 v0.0.0-20180517000239-4f5cfb03fabf // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/faiface/mainthread v0.0.0-20171120011319-8b78f0a41ae3 // indirect
+	github.com/fatih/structs v1.1.0 // indirect
+	github.com/go-gl/gl v0.0.0-20211210172815-726fda9656d6 // indirect
+	github.com/gogo/protobuf v1.3.1 // indirect
+	github.com/google/go-cmp v0.5.6 // indirect
+	github.com/imdario/mergo v0.3.9 // indirect
+	github.com/imroc/req v0.3.0 // indirect
+	github.com/mattn/go-sqlite3 v1.14.1 // indirect
+	github.com/miekg/dns v1.1.27 // indirect
+	github.com/pkg/browser v0.0.0-20210911075715-681adbf594b8 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/sahilm/fuzzy v0.1.0 // indirect
+	github.com/stretchr/objx v0.1.0 // indirect
+	golang.org/x/crypto v0.0.0-20191011191535-87dc89f01550 // indirect
+	golang.org/x/image v0.0.0-20220302094943-723b81ca9867 // indirect
+	golang.org/x/net v0.0.0-20200324143707-d3edc9973b7e // indirect
+	golang.org/x/sys v0.0.0-20220315194320-039c03cc5b86 // indirect
+	gopkg.in/eapache/queue.v1 v1.1.0 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )
